+++ srcs/robot/robot_lifecycle.h
@@ -2,6 +2,6 @@
 
 #include "types.h"
 
-void initializeCoverageRobot(Robot &rb, int maxEnergy);
+void initializeCoverageRobot(Robot &rb, double maxEnergy);
 
 void rechargeRobot(Robot &rb);
