+++ srcs/robot/tactical_yield.h
@@ -1,20 +1,19 @@
-
 #pragma once
 
 #include "types.h"
 
 struct TacticalYieldConfig
 {
-    int maxCandidateCost = 3;
+    double maxCandidateCost = 3.0;
 };
 
 struct TacticalYieldResult
 {
     bool found = false;
     Cell target = {0, 0};
-    int costFromRobot = INF;
-    int costToBase = INF;
-    int score = INF;
+    double costFromRobot = INF;
+    double costToBase = INF;
+    double score = INF;
 };
 
 TacticalYieldResult findTacticalYieldCell(
